[REFACTOR/#220] Profile 폴더링
