+++ app/src/main/AndroidManifest.xml
@@ -145,12 +145,12 @@
             android:screenOrientation="portrait" />
 
         <activity
-            android:name="com.going.presentation.profile.ProfileActivity"
+            android:name="com.going.presentation.profile.my.ProfileActivity"
             android:exported="false"
             android:screenOrientation="portrait" />
 
         <activity
-            android:name="com.going.presentation.profile.ProfileEditActivity"
+            android:name="com.going.presentation.profile.edit.ProfileEditActivity"
             android:exported="false"
             android:screenOrientation="portrait" />
